генерация разметки offers
