+++ src/main.js
@@ -1,6 +1,6 @@
 import {createTripInfoTemplate} from "./components/TripInfoTemplate.js";
-import {createEventEditTemplate} from "./components/EventEditTemplate.js";
-import {createTripEvent} from "./components/TripEvent.js";
+import {createEventEditTemplate} from "./components/event-edit.js";
+import {createTripEvent} from "./components/event.js";
 import {createTripFiltersTemplate} from "./components/TripFiltersTemplate.js";
 import {createTripSort} from "./components/TripSort.js";
 import {createTripTabs} from "./components/TripTabs.js";
